+++ kubetools/cli/deploy.py
@@ -259,6 +259,7 @@ def remove(ctx, yes, force, do_cleanup, namespace, app_or_project_names):
 def cleanup(ctx, yes, namespace):
     '''
     Cleans up a namespace by removing orphaned objects.
+    Will delete the namespace if it's empty after cleanup.
     '''
 
     build = Build(
@@ -266,14 +267,14 @@ def cleanup(ctx, yes, namespace):
         namespace=namespace,
     )
 
-    replica_sets_to_delete, pods_to_delete = get_cleanup_objects(build)
+    namespace_to_delete, replica_sets_to_delete, pods_to_delete = get_cleanup_objects(build)
 
-    if not any((replica_sets_to_delete, pods_to_delete)):
+    if not any((namespace_to_delete, replica_sets_to_delete, pods_to_delete)):
         click.echo('Nothing to do 👍!')
         return
 
     log_cleanup_changes(
-        build, replica_sets_to_delete, pods_to_delete,
+        build, namespace_to_delete, replica_sets_to_delete, pods_to_delete,
         message='Executing changes:' if yes else 'Proposed changes:',
         name_formatter=lambda name: click.style(name, bold=True),
     )
@@ -286,6 +287,7 @@ def cleanup(ctx, yes, namespace):
 
     execute_cleanup(
         build,
+        namespace_to_delete,
         replica_sets_to_delete,
         pods_to_delete,
     )