+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v11.2
+- Implicitly create target namespace if it does not exist
+- The `--cleanup` flag will now remove remaining empty namespaces
+- Replace `yaml.load` with `yaml.safe_load` to avoid CLI warning
+
 # v11.1.1
 - Fix `-f` / `--file` handling NoneType attribute error (introduced in v11.1)